requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1
PIL
torchvision
torchvision
argparse
matplotlib.pyplot
numpy
sklearn
torchvision
tensorboard
